refact(OAuth): Update with latest API and refactor code
